+++ example/lib/components/demo/demo_result_component.dart
@@ -10,12 +10,12 @@ import 'package:flutter/material.dart';
 
 class DemoResultComponent extends StatefulWidget {
   @override
-  _DemoResultComponentState createState() => new _DemoResultComponentState();
+  _DemoResultComponentState createState() => _DemoResultComponentState();
 }
 
 class _DemoResultComponentState extends State<DemoResultComponent> {
   @override
   Widget build(BuildContext context) {
-    return new Container();
+    return Container();
   }
 }